backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis (caching)
redis==5.2.1

# HTTP / Crawling
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# Scoring utilities
python-whois==0.9.5
textstat==0.7.4

# AI APIs
anthropic==0.43.0

# Email
resend==2.5.0

# Templates
jinja2==3.1.5

# CORS
python-multipart==0.0.20
